Update jarvis_AI.py
Speech recognition system
A User will turn on the system and will give a voice command. The system will recognize the given voice command and based on that voice command, it will perform the operation. For example, users can say: “Open Google” and if the browser is installed, it will open the Google search engine otherwise it will handle the exception or can give proper suggestions.